Create dataGraphUI.py
I coded a detailed User Interface for a software project. The software's purpose is to assist with an experiment by tracking the data of three variables: Temperature, pH, and Flow Rate. 
The UI offers a timer widget that starts the collection of data and displays it on the graph widgets live. The user can input specific target values on the side input lines for each variable that will be used for tracking in the tracker widgets (below the graph tab), which display the quality status of the current data. Currently, data is randomly generated and plotted on the graph for testing purposes.